use int32 instead uint64 for scrolling buffer indexing
binary search does not work correctly if index becomes negative
and search should be stopped but with uint it rolls over and
overindexes

+++ debug_gui.h
@@ -41,7 +41,7 @@ struct Scalar {
 
     void startBuffering() {
         if (buffer == nullptr) {
-            buffer = std::make_unique<ScrollingBuffer>(size_t(1e6));
+            buffer = std::make_unique<ScrollingBuffer>(int32_t(1e6));
         }
     }
